Move Formatter to its own file
